- 스탯 올리는 장착템 소유한 경우 주사위 굴리고 나서 하나 더 추가되는 현상 수정 - 주사위 판 스탯 언어적용 - 시간 수정 - clock 사운드 교체 - 사운드 파일 다 모노로 변경 - 메인화면 텍스처 교체

+++ Assets/ToBeFree/Scripts/Action.cs
@@ -300,7 +300,7 @@ namespace ToBeFree
 				yield break;
 			}
 
-			int moveTimePerCity = TimeTable.Instance.MoveTimePerAction;
+			float moveTimePerCity = TimeTable.Instance.MoveTimePerAction;
 
 			if (actionName == eEventAction.MOVE_BUS)
 			{
